requirements.txt
chatbot==1.5.2b
langchain==0.3.9
langchain_openai==0.2.10
pinecone==5.4.1
python-dotenv==1.0.1
google-api-python-client
protoc_gen_openapiv2
grpcio
